microservices/5001-stock-predict-core/requirements.txt
grpcio
grpcio-tools
tensorflow
numpy
scikit-learn
google-cloud-bigquery
boto3
pandas
db-dtypes
